requirements.txt
discord.py==2.4.0
asyncio
flask
python-dotenv
Pillow
